FS: acomodo los tipos de resultados para que sean uint8

+++ LFS/API.h
@@ -2,16 +2,16 @@
 #ifndef LFS_API_h__
 #define LFS_API_h__
 
-#include <Console.h>
-#include <string.h>
-#include <libcommons/string.h>
-#include "LissandraLibrary.h"
 #include "FileSystem.h"
+#include "LissandraLibrary.h"
+#include <Console.h>
+#include <libcommons/string.h>
+#include <string.h>
 
 void select_api(char const* nombreTabla, uint16_t key);
-int insert(char const* nombreTabla, uint16_t key, char const* value, uint64_t timestamp);
-int create(char const* nombreTabla, uint8_t tipoConsistencia, uint16_t numeroParticiones, uint32_t compactionTime);
+uint8_t insert(char const* nombreTabla, uint16_t key, char const* value, uint64_t timestamp);
+uint8_t create(char const* nombreTabla, uint8_t tipoConsistencia, uint16_t numeroParticiones, uint32_t compactionTime);
 void* describe(char const* nombreTabla);
-int drop(char const* nombreTabla);
+uint8_t drop(char const* nombreTabla);
 
 #endif //LFS_API_h__
